+++ example/main.rs
@@ -1,3 +1,5 @@
+use std::fs::OpenOptions;
+
 use dbgout::{Color, debug, set_debug_color, set_debug_writer};
 
 fn do_debug_prints(prestr: &str) {
@@ -12,4 +14,23 @@ fn do_debug_prints(prestr: &str) {
     debug!(true, "{} - this will always print", prestr);
 }
 
-fn main() {}
+fn main() -> Result<(), Box<dyn std::error::Error>> {
+    // just a regular debug print
+    do_debug_prints("regular");
+
+    eprintln!();
+
+    // set the colour
+    set_debug_color(Color::Blue);
+    do_debug_prints("blue");
+
+    // send to a file
+    let file = OpenOptions::new()
+        .create(true)
+        .append(true)
+        .open("debug.log")?;
+    set_debug_writer(Box::new(file));
+    do_debug_prints("file");
+
+    Ok(())
+}