+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changlog
 
+## [2.2.0] - 2026-03-23
+### Changed
+- `debug!` now supports calling without any format string
+
 ## [2.1.0] - 2026-03-23
 ### Fixed
 - `debug!` matching `$debug_mode:literal` on non-boolean literals